# Initial quote perf tuning by increasing the distribution percent from 5% to 10%