--- a/src/routers/alpha-router/config.ts
+++ b/src/routers/alpha-router/config.ts
@@ -6,6 +6,30 @@ export const DEFAULT_ROUTING_CONFIG_BY_CHAIN = (
   chainId: ChainId
 ): AlphaRouterConfig => {
   switch (chainId) {
+    case ChainId.MAINNET:
+      return {
+        v2PoolSelection: {
+          topN: 3,
+          topNDirectSwaps: 1,
+          topNTokenInOut: 5,
+          topNSecondHop: 2,
+          topNWithEachBaseToken: 2,
+          topNWithBaseToken: 6,
+        },
+        v3PoolSelection: {
+          topN: 2,
+          topNDirectSwaps: 2,
+          topNTokenInOut: 3,
+          topNSecondHop: 1,
+          topNWithEachBaseToken: 3,
+          topNWithBaseToken: 5,
+        },
+        maxSwapsPerPath: 3,
+        minSplits: 1,
+        maxSplits: 7,
+        distributionPercent: 20,
+        forceCrossProtocol: false,
+      };
     // Optimism
     case ChainId.OPTIMISM:
     case ChainId.OPTIMISM_GOERLI: