--- a/src/routers/alpha-router/config.ts
+++ b/src/routers/alpha-router/config.ts
@@ -27,7 +27,7 @@ export const DEFAULT_ROUTING_CONFIG_BY_CHAIN = (
         maxSwapsPerPath: 3,
         minSplits: 1,
         maxSplits: 7,
-        distributionPercent: 20,
+        distributionPercent: 10,
         forceCrossProtocol: false,
       };
     // Optimism