--- a/src/routers/alpha-router/config.ts
+++ b/src/routers/alpha-router/config.ts
@@ -67,25 +67,25 @@ export const DEFAULT_ROUTING_CONFIG_BY_CHAIN = (
     default:
       return {
         v2PoolSelection: {
-          topN: 0,
-          topNDirectSwaps: 2,
-          topNTokenInOut: 2,
-          topNSecondHop: 0,
-          topNWithEachBaseToken: 0,
-          topNWithBaseToken: 0,
+          topN: 3,
+          topNDirectSwaps: 1,
+          topNTokenInOut: 5,
+          topNSecondHop: 2,
+          topNWithEachBaseToken: 2,
+          topNWithBaseToken: 6,
         },
         v3PoolSelection: {
-          topN: 0,
+          topN: 2,
           topNDirectSwaps: 2,
-          topNTokenInOut: 2,
-          topNSecondHop: 0,
-          topNWithEachBaseToken: 0,
-          topNWithBaseToken: 0,
+          topNTokenInOut: 3,
+          topNSecondHop: 1,
+          topNWithEachBaseToken: 3,
+          topNWithBaseToken: 5,
         },
         maxSwapsPerPath: 2,
         minSplits: 1,
         maxSplits: 2,
-        distributionPercent: 20,
+        distributionPercent: 100,
         forceCrossProtocol: false,
       };
   }