- **What kind of change does this PR introduce?** (Bug fix, feature, docs update, ...)
Perf tuning

- **What is the current behavior?** (You can also link to an open issue here)
Initial perf tuning mainnet quote latency.

https://github.com/Uniswap/smart-order-router/pull/306 - is the no-op PR, with
[mainnet alpha exactIn](https://github.com/Uniswap/smart-order-router/actions/runs/5860226437/job/15887849620?pr=306#logs) running 15m 45s
[mainnet alpha exactOut](https://github.com/Uniswap/smart-order-router/actions/runs/5860226437/job/15887849813?pr=306) running 14m 50s
[mixed routes](https://github.com/Uniswap/smart-order-router/actions/runs/5860226437/job/15887850004?pr=306) running 6m 43s

- **What is the new behavior (if this is a feature change)?**
https://github.com/Uniswap/smart-order-router/pull/307 - changing distributionPercent  from 5 to 10 for mainnet only
[mainnet alpha exactInt](https://github.com/Uniswap/smart-order-router/actions/runs/5874134666/job/15928424532?pr=307) running 6m 57s
[mainnet alpha exactOut](https://github.com/Uniswap/smart-order-router/actions/runs/5874134666/job/15928424629?pr=307) running 10m 47s
[mixed routes](https://github.com/Uniswap/smart-order-router/actions/runs/5874134666/job/15928424794?pr=307) running 2m 23s

- **Other information**:
`DistributionPercent` seems in rough inverse proportion to the quote latency.

Also tried the following
https://github.com/Uniswap/smart-order-router/pull/307 - changing distributionPercent  from 5 to 20 for mainnet only
[mainnet alpha exactInt](https://github.com/Uniswap/smart-order-router/actions/runs/5873989343/job/15928054833?pr=307) running 3m 51s
[mainnet alpha exactOut](https://github.com/Uniswap/smart-order-router/actions/runs/5873989343/job/15928054552?pr=307) running 4m 39s
[mixed routes](https://github.com/Uniswap/smart-order-router/actions/runs/5873989343/job/15928054720?pr=307) running 1m 22s