@jsy1218 2023-08-25
We can collect data from api endpoint now, no longer need this experimental PR.